requirements.txt
flask
flask-cors
python-telegram-bot
requests
sqlite3